requirements.txt
streamlit
sqlalchemy
mysql-connector-python
python-dotenv
pandas
requests
schedule
altair
